refactor: restructure login and dev pages for improved mobile compatibility and styling
Also add glow to the (non)DraggableEvents for whatever reason
