--- src/pages/DevOld.css
@@ -1,33 +1,0 @@
-.grid-container {
-    display: grid;
-    grid-template-columns: 40% 60%;
-    grid-gap: 2em;
-}
-
-.left-col {
-    padding-left: 2em;
-}
-
-.student-date {
-    display: flex;
-    justify-content: space-evenly;
-    padding-top: 2em;
-    padding-bottom: 4em;
-}
-
-.errors {
-    justify-self: right;
-    padding-right: 2em;
-    padding-top: 2em;
-}
-
-.date-input {
-    padding-bottom: 2em;
-    background-color: aqua;
-}
-
-html,
-body {
-    height: 100%;
-    width: 100%;
-}
