--- src/pages/LoginPage.css
@@ -1,9 +1,0 @@
-.login-page-container {
-    margin: 10px 100px 10px 100px;
-    font-size: 18px;
-}
-
-.login-page-container h4 {
-    color: white;
-    font-size: 2rem;
-}
